+++ Cargo.toml
@@ -1,15 +1,24 @@
 [package]
-name = "glob-match"
+name = "fast-glob"
+version = "0.1.0"
 description = "An extremely fast glob matcher"
-version = "0.2.1"
+repository = "https://github.com/shulaoda/fast-glob.git"
+authors = [
+    "Devon Govett <devongovett@gmail.com>",
+    "shulaoda <laipichan@qq.com>",
+]
 license = "MIT"
-authors = ["Devon Govett <devongovett@gmail.com>"]
 edition = "2021"
 
 [dev-dependencies]
-criterion = "0.4.0"
+criterion = "0.5.1"
 glob = "0.3.1"
-globset = "0.4.10"
+globset = "0.4.14"
+
+[[test]]
+name = "test"
+path = "tests/test.rs"
+
 
 [[bench]]
 name = "bench"
