+++ benches/bench.rs
@@ -1,8 +1,8 @@
+use fast_glob::glob_match;
 use criterion::{criterion_group, criterion_main, Criterion};
-use glob_match::*;
 
-const PATH: &'static str = "some/a/bigger/path/to/the/crazy/needle.txt";
 const GLOB: &'static str = "some/**/needle.txt";
+const PATH: &'static str = "some/a/bigger/path/to/the/crazy/needle.txt";
 
 #[inline]
 fn glob(pat: &str, s: &str) -> bool {
@@ -16,7 +16,7 @@ fn globset(pat: &str, s: &str) -> bool {
   pat.is_match(s)
 }
 
-fn glob_match_crate(b: &mut Criterion) {
+fn mine_crate(b: &mut Criterion) {
   b.bench_function("mine", |b| b.iter(|| assert!(glob_match(GLOB, PATH))));
 }
 
@@ -28,5 +28,5 @@ fn globset_crate(b: &mut Criterion) {
   b.bench_function("globset_crate", |b| b.iter(|| assert!(globset(GLOB, PATH))));
 }
 
-criterion_group!(benches, globset_crate, glob_crate, glob_match_crate);
+criterion_group!(benches, globset_crate, glob_crate, mine_crate);
 criterion_main!(benches);
