--- fuzz/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "glob-match-fuzz"
-version = "0.0.0"
-publish = false
-edition = "2021"
-
-[package.metadata]
-cargo-fuzz = true
-
-[dependencies]
-libfuzzer-sys = "0.4"
-arbitrary = { version = "1", features = ["derive"] }
-
-[dependencies.glob-match]
-path = ".."
-
-# Prevent this from interfering with workspaces
-[workspace]
-members = ["."]
-
-[profile.release]
-debug = 1
-
-[lib]
-name = "fuzz_local"
-path = "fuzz_targets/lib.rs"
-crate-types = ["rlib"]
-
-[[bin]]
-name = "both_fuzz"
-path = "fuzz_targets/both_fuzz.rs"
-test = false
-doc = false
-
-[[bin]]
-name = "pattern_on_itself"
-path = "fuzz_targets/pattern_on_itself.rs"
-test = false
-doc = false
