--- fuzz/fuzz_targets/lib.rs
@@ -1,7 +1,0 @@
-use arbitrary::Arbitrary;
-
-#[derive(Debug, Arbitrary)]
-pub struct Data<'a> {
-  pub pat: &'a str,
-  pub input: &'a str,
-}
